Change implementation to abstract actor class
Hunter and Animal classes extend the Actor class that has much of the
functionality of the animal class. For simplicity, the actor class
retains alive and dead and a Hunter is 'killed' when they become
inactive.

Change the Simulator class to implement Hunter with a yellow box and
create them at the beginning of the simulation.

Change the SimulatorTest class to wait in between every 5 steps. This
allows for a better appreciation of the system.